refactor: move auth to core
