+++ pkg/core/auth/login/static/mas-sso-redirect-page.html
