+++ pkg/core/auth/login/static/sso-redirect-page.html
